DMDc no parametrization
