+++ framework/Models/ROM.py
@@ -266,7 +266,7 @@ class ROM(Dummy):
     inputSpecification.addSub(InputData.parameterInputFactory("exactModes", contentType=InputTypes.BoolType))
     inputSpecification.addSub(InputData.parameterInputFactory("optimized", contentType=InputTypes.BoolType))
     inputSpecification.addSub(InputData.parameterInputFactory("dmdType", contentType=InputTypes.StringType))
-    inputSpecification.addSub(InputData.parameterInputFactory("Actuators", contentType=InputTypes.StringType))
+    inputSpecification.addSub(InputData.parameterInputFactory("Actuators", contentType=InputTypes.StringListType))
     inputSpecification.addSub(InputData.parameterInputFactory("SubtractNormUXY", contentType=InputTypes.BoolType))
     # for deep learning neural network
     #inputSpecification.addSub(InputData.parameterInputFactory("DNN", InputTypes.StringType))
